refactor: database -> db
